weaver_service/requirements.txt
boto3==1.36.26
huggingface_hub==0.28.1
numpy==2.2.2
Pillow==11.1.0
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1
# For real CatVTON mode, install matching CUDA build of torch in your environment/container.
# torch==2.5.1
